Initial commit Command line semantics changed:   -every = loads all servers found   -name = filter servers by name   -host = filter servers by default-host
Basically each parameter filters the list and could return multiple servers.
Multiple servers are only started if the -every option is present.

This means that if only one server is defined in the database, you do not
need to specify any options. You can have multiple servers with the same
name or default-host (but not uuid).